Finish the changes to UI/UX for features #221
*finish UI/UX for adding and editing features
*remove edit component and replace by dialog component
