+++ phenopacketlab-frontend/src/app/component/phenopacket/phenotypic-feature/phenotypic-feature.component.html
@@ -1,21 +1,23 @@
 <div class="grid p-fluid">
   <div class="field col-12">
-    <app-search-box
-      [showHint]="false"
-      [itemName]="'Phenotypic feature'"
-      [searchLabel]="'Phenotypic feature'"
-      [placeHolderTxt]="'Enter a phenotypic feature name'"
-      [searchType]="'feature'"
-      (selectedItemChange)="featureItemSelected($event)"
-    ></app-search-box>
-    
-  </div>
-  <div class="field col-12">
     <p-table
       [value]="phenotypicFeatures"
       dataKey="key"
       styleClass="p-datatable-sm"
     >
+    <ng-template pTemplate="caption">
+      <div class="flex align-items-center justify-content-between">
+        Added phenotypic features
+        <button
+          pButton
+          icon="pi pi-plus"
+          label="Add phenotypic feature"
+          class="p-button-text"
+          style="width: 250px"
+          (click)="addPhenotypicFeature()"
+        ></button>
+      </div>
+    </ng-template>
       <ng-template pTemplate="header">
         <tr>
           <th style="width: 5rem"></th>
