+++ phenopacketlab-frontend/src/app/component/shared/dialog/phenotypic-feature-search-dialog/phenotypic-feature-search-dialog.component.html
@@ -1,9 +1,3 @@
-<!-- <app-phenotypic-feature-edit
-  [interpretation]="interpretation"
-  [phenopacket]="phenopacket"
-  (interpretationChange)="updateInterpretation($event)"
-  [profile]="profile"
-></app-phenotypic-feature-edit> -->
 <div class="grid p-fluid">
   <div class="field col-12">
     <p-tabView>
