--- phenopacketlab-frontend/src/app/component/shared/edit/phenotypic-feature-edit/phenotypic-feature-edit.component.scss
@@ -1,88 +1,0 @@
-h2 {
-    font-size: 20px;
- }
-   
- h3 {
-    font-size: 15px;
- }
-   
- h2, h3 {
-    width: 100%;
-    height: 60px;
-    margin: 0;
-    padding: 0;
-    display: inline;
- }
- 
- .margin {
-     margin: 10px;
- }
- .status-radio-group {
-     display: flex;
-     flex-direction: column;
-     margin: 15px 10px;
- }
- .status-radio-button {
-     margin: 5px;
- }
- .half-width {
-     width: 50%;
- }
- 
- /* .example-form {
-   min-width: 150px;
-   max-width: 500px;
-   width: 100%;
- } */
- 
- .example-full-width {
-     width: 100%;
-   }
-   
-   
-   .example-container {
-     height: 800px;
-     overflow: auto;
-   }
-   
-   /* .mat-table {
-     overflow: auto;
-     width: 100%;
-     max-height: 100px;
-     table-layout: fixed;
-   } */
-   
-   /* th, td {
-     overflow: hidden;
-     width: 60px;
-     text-overflow: ellipsis;
-     white-space: nowrap;
-   } */
-   
-   
-   .example-element-detail {
-     overflow: hidden;
-     display: flex;
-   }
-   
-   .example-element-row td {
-     border-bottom-width: 0;
-   }
-   
-   tr.example-detail-row {
-     height: 0;
-   }
-   
-   ::ng-deep .mat-tooltip {
-     /* your own custom styles here */
-     /* e.g. */
-   
-   
-     background: #616161e6;
-     font-size: small;
-     color: white !important;
-   }
-   
-   .cardSearchBox {
-     background: #fafafa;
-   }
